1.更新编译为ff4.0--ijk0.8.8--20210426--001版本; 2.Android 11以上高版本报错问题; 3.音量、光亮、进度调节逻辑修改;

+++ app/src/main/java/androidx/ijk/helper/OnIjkVideoTouchListener.java
@@ -23,7 +23,7 @@ public interface OnIjkVideoTouchListener {
      * @param value   音量值
      * @param percent 音量百分比[0-1]
      */
-    void onVideoChangeVoice(float value, float percent);
+    void onVideoChangeVoice(int value, float percent);
 
     /**
      * 改变进度开始
